requirements.txt
pytorch-lightning==1.7.7
torchmetrics==0.11.4
matplotlib
opencv-python
tqdm
imageio
path
scipy
configargparse
kornia
